requirements.txt
google-cloud-texttospeech
streamlit
python-dotenv
